Correção

+++ Lista 3 em c/mediavetor_extra.c
@@ -3,14 +3,15 @@
 
 int main() {
 
-    double notas[QTD];
-    double media = 0.0;
+    float notas[QTD];
+    float media = 0.0;
     
     for (int i = 0; i < QTD; i++) {
         printf("Digite a nota: ");
-        scanf("%lf", &notas[i]);
+        scanf("%f", &notas[i]);
         media += notas[i];
     }
-    printf("A média entre eles será: %.2lf\n", media / QTD);
+    
+    printf("A média entre eles será: %.2f\n", media / QTD);
     return 0;
 }
